Allow for audio files to be replayed
Files could not be replayed earlier because of a bug/feature,
where calling the `aud->stop()` function made it so that
audio files could not be replayed anymore.
This change fixes the issue by recreating the out, aud, and clip values
every time in a separate `playClip()` function.

This commit also moves things around and splits functionality
into different functions for cleanliness.